<feat>(mainwindow): add power limiting functionality with UI controls
- Add power limit spin box with apply/cancel buttons and percentage display
- Implement power limiting logic with visual feedback using color-coded labels
- Add connections for power limit button clicks and value changes
- Create methods to calculate and update system power limits
- Update log messages with emoji indicators for better user experience

<refactor>(topology): replace internal power calculation with external API calls

- Replace get_totalpower() with external ::get_system_gross_power() function
- Add restrictPower method to handle power limitation logic
- Update requestPower to use external power limit functions
- Remove internal power calculation implementation

<docs>(about): clarify software purpose for educational and research use

- Update description to specify electrical system design feasibility validation
- Emphasize non-commercial usage restrictions

<chore>(ui): improve layout and styling for power control interface

- Reorganize power control widgets into horizontal layouts
- Add size policies for better responsive behavior
- Style power limit percentage label with custom CSS
- Add prefix to priority spin box for better UX
```

+++ pwralloc/pau_tactic.h
@@ -39,7 +39,7 @@ enum METABOLIN
 // union PCU_RawData retriver_PCU_RawData(ID_TYPE id, PCURawData data_type);
 ID_TYPE get_neighbor_lower_alpha(ID_TYPE nodeid);
 ID_TYPE get_neighbor_lower_beta(ID_TYPE nodeid);
-size_t get_plug_chargingnodes_cnt(ID_TYPE plugid);
+int get_plug_chargingnodes_cnt(ID_TYPE plugid);
 size_t makeScore(enum Senario senario, int quota, ID_TYPE plugid, ID_TYPE neighbor_plugid, ID_TYPE nodeid, ID_TYPE neighbor_nodeid);
 size_t get_plug_requiredPower(ID_TYPE plugid);
 PRIOR get_plug_priority(ID_TYPE plugid);
@@ -63,7 +63,7 @@ bool get_plug_sequent_flag(ID_TYPE plugid);
 void set_plug_sequent_flag(ID_TYPE plugid, bool val);
 
 PRIOR get_node_priority(ID_TYPE nodeid);
-size_t get_node_module_cnt(ID_TYPE nodeid);
+int get_node_module_cnt(ID_TYPE nodeid);
 size_t get_allover_modules_cnt(void);
 
 int get_node_available_power(ID_TYPE nodeid);
@@ -76,6 +76,7 @@ bool is_node_pseudocycledon(ID_TYPE nodeid);
 void recover_node_pseudocycledon(ID_TYPE plugid, ID_TYPE nodeid);
 bool have_plug_occupied_matrixnode(ID_TYPE plugid);
 enum METABOLIN metabole_alethes(unsigned char nodeid, unsigned char relayid, FlowMap *pflow_map);
+
 #ifdef __cplusplus
 extern "C"
 {
